+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/request/balance/BalanceChargeRequest.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.request;
+package kr.hhplus.be.server.reservation.interfaces.web.dto.request.balance;
 
 public record BalanceChargeRequest(
     Long walletId,
