+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/balance/BalanceChargeResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.response;
+package kr.hhplus.be.server.reservation.interfaces.web.dto.response.balance;
 
 public record BalanceChargeResponse(
         long walletId,
