+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/reservation/AvailableDatesResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.response;
+package kr.hhplus.be.server.reservation.interfaces.web.dto.response.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
