+++ src/main/java/kr/hhplus/be/server/reservation/interfaces/web/dto/response/reservation/AvailableSeatsResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.interfaces.web.dto.response;
+package kr.hhplus.be.server.reservation.interfaces.web.dto.response.reservation;
 
 import java.util.List;
 import kr.hhplus.be.server.reservation.infrastructure.persistence.seat.SeatEntity;
